Rename: macros.h -> logger.h
