+++ lib/helpers.cxx
@@ -23,7 +23,7 @@
 #include <fstream>
 #include <stdexcept>
 
-#include "macros.h"
+#include "logger.h"
 
 
 //----------------------------------------------------------------------------
